requirements.txt
GitPython==0.3.2.RC1
MySQL-python==1.2.5
mosquitto==1.2.3
peewee==2.2.3
requests==2.2.1

